Bug Fix: replacing list instead of appending
During the processing for the a-level route, the for loop was originally
overwriting the lists with the values of the decoded data. This had the
effect of replacing the lists with string of the last packet's headers
and data (i.e. just an ACK), instead of the intended functionality of
having a list for each header type throughout the transaction.